+++ index.es6.js
@@ -1,4 +1,0 @@
-export { default } from './es6/index'
-export { default as gradation, day, month, year } from './es6/gradation'
-export { intl_supported_locale as intlSupportedLocale } from './es6/locale'
-export { default as RelativeTimeFormat } from './es6/RelativeTimeFormat'