+++ index.js
@@ -1,0 +1,3 @@
+export { default } from './modules/JavascriptTimeAgo'
+export { intlSupportedLocale as intlSupportedLocale } from './modules/locale'
+export { default as RelativeTimeFormat } from './modules/RelativeTimeFormat'